Merge from original repo

+++ ydb/core/blobstorage/pdisk/mock/pdisk_mock.cpp
@@ -3,6 +3,8 @@
 #include <ydb/core/util/stlog.h>
 #include <ydb/core/util/interval_set.h>
 
+#include <ydb/core/blobstorage/pdisk/blobstorage_pdisk_util_space_color.h>
+
 namespace NKikimr {
 
 #ifdef _MSC_VER
@@ -41,6 +43,7 @@ struct TPDiskMockState::TImpl {
     ui32 NextFreeChunk = 1;
     std::unordered_map<TString, ui32> Blocks;
     TIntervalSet<ui64> Corrupted;
+    NPDisk::TStatusFlags StatusFlags;
 
     TImpl(ui32 nodeId, ui32 pdiskId, ui64 pdiskGuid, ui64 size, ui32 chunkSize)
         : NodeId(nodeId)
@@ -51,6 +54,7 @@ struct TPDiskMockState::TImpl {
         , TotalChunks(Size / ChunkSize)
         , AppendBlockSize(4096)
         , NextFreeChunk(1)
+        , StatusFlags(NPDisk::TStatusFlags{})
     {}
 
     TImpl(const TImpl&) = default;
@@ -147,9 +151,19 @@ struct TPDiskMockState::TImpl {
     void DeleteChunk(TOwner& owner, TChunkIdx chunkIdx) {
         const ui32 num = owner.ReservedChunks.erase(chunkIdx) + owner.CommittedChunks.erase(chunkIdx);
         Y_VERIFY(num);
+        owner.ChunkData.erase(chunkIdx);
         const bool inserted = FreeChunks.insert(chunkIdx).second;
         Y_VERIFY(inserted);
         AdjustFreeChunks();
+    }
+
+    void UncommitChunk(TOwner& owner, TChunkIdx chunkIdx) {
+        if (owner.ReservedChunks.contains(chunkIdx)) {
+        } else if (owner.CommittedChunks.erase(chunkIdx)) {
+            owner.ReservedChunks.insert(chunkIdx);
+        } else {
+            Y_FAIL();
+        }
     }
 
     void SetCorruptedArea(ui32 chunkIdx, ui32 begin, ui32 end, bool enabled) {
@@ -200,6 +214,14 @@ struct TPDiskMockState::TImpl {
             }
         }
     }
+
+    void SetStatusFlags(NPDisk::TStatusFlags flags) {
+        StatusFlags = flags;
+    }
+
+    void SetStatusFlags(NKikimrBlobStorage::TPDiskSpaceColor::E spaceColor) {
+        StatusFlags = SpaceColorToStatusFlag(spaceColor);
+    }
 };
 
 TPDiskMockState::TPDiskMockState(ui32 nodeId, ui32 pdiskId, ui64 pdiskGuid, ui64 size, ui32 chunkSize)
@@ -231,6 +253,14 @@ TIntervalSet<i64> TPDiskMockState::GetWrittenAreas(ui32 chunkIdx) const {
 
 void TPDiskMockState::TrimQuery() {
     Impl->TrimQuery();
+}
+
+void TPDiskMockState::SetStatusFlags(NKikimrBlobStorage::TPDiskSpaceColor::E spaceColor) {
+    Impl->SetStatusFlags(spaceColor);
+}
+
+void TPDiskMockState::SetStatusFlags(NPDisk::TStatusFlags flags) {
+    Impl->SetStatusFlags(flags);
 }
 
 TPDiskMockState::TPtr TPDiskMockState::Snapshot() {
@@ -445,7 +475,11 @@ public:
                         Impl.CommitChunk(owner, chunk);
                     }
                     for (const TChunkIdx chunk : cr.DeleteChunks) {
-                        Impl.DeleteChunk(owner, chunk);
+                        if (cr.DeleteToDecommitted) {
+                            Impl.UncommitChunk(owner, chunk);
+                        } else {
+                            Impl.DeleteChunk(owner, chunk);
+                        }
                     }
                     isStartingPoint = cr.IsStartingPoint;
                 }
@@ -484,6 +518,28 @@ public:
                 Send(owner.CutLogId, new NPDisk::TEvCutLog(ownerId, owner.OwnerRound, lsn, 0, 0, 0, 0));
             }
         }
+    }
+
+    void Handle(NPDisk::TEvChunkForget::TPtr ev) {
+        auto *msg = ev->Get();
+        NKikimrProto::EReplyStatus status = NKikimrProto::OK;
+        TString errorReason;
+        if (const auto it = Impl.Owners.find(msg->Owner); it == Impl.Owners.end()) {
+            Y_FAIL("invalid Owner");
+        } else if (it->second.Slain) {
+            status = NKikimrProto::INVALID_OWNER;
+            errorReason = "VDisk is slain";
+        } else if (msg->OwnerRound != it->second.OwnerRound) {
+            status = NKikimrProto::INVALID_ROUND;
+            errorReason = "invalid OwnerRound";
+        } else {
+            TImpl::TOwner& owner = it->second;
+            PDISK_MOCK_LOG(DEBUG, PDMxx, "received TEvChunkForget", (Msg, msg->ToString()), (VDiskId, owner.VDiskId));
+            for (const TChunkIdx chunkIdx : msg->ForgetChunks) {
+                Impl.DeleteChunk(owner, chunkIdx);
+            }
+        }
+        Send(ev->Sender, new NPDisk::TEvChunkForgetResult(status, {}, errorReason), 0, ev->Cookie);
     }
 
     void Handle(NPDisk::TEvReadLog::TPtr ev) {
@@ -671,12 +727,13 @@ public:
     }
 
     NPDisk::TStatusFlags GetStatusFlags() {
-        return {};
+        return Impl.StatusFlags;
     }
 
     STRICT_STFUNC(StateFunc,
         hFunc(NPDisk::TEvYardInit, Handle);
         hFunc(NPDisk::TEvLog, Handle);
+        hFunc(NPDisk::TEvChunkForget, Handle);
         hFunc(NPDisk::TEvMultiLog, Handle);
         cFunc(EvResume, HandleLogQ);
         hFunc(NPDisk::TEvReadLog, Handle);
